fix: 멤버 추가 리스트 보여주는 기능 수정
